+++ Frends.Salesforce.DeleteSObject/Frends.Salesforce.DeleteSObject/DeleteSObject.cs
@@ -7,6 +7,7 @@ using System.Net.Http;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
+using Frends.Salesforce.DeleteSObject.Helpers;
 
 [assembly: InternalsVisibleTo("Frends.Salesforce.DeleteSObject.Tests")]
 
@@ -27,49 +28,30 @@ public class Salesforce
     /// <returns>Object { object Body, bool RequestIsSuccessful, Exception ErrorException, string ErrorMessage, string Token }</returns>
     public static async Task<Result> DeleteSObject(
         [PropertyTab] Input input,
-        [PropertyTab] Options options,
+        [PropertyTab] Connection connection,
         CancellationToken cancellationToken
     )
     {
-        var accessToken = string.Empty;
 
         try
         {
-            if (string.IsNullOrWhiteSpace(input.Domain)) throw new ArgumentNullException("Domain cannot be empty.");
-            if (string.IsNullOrWhiteSpace(input.SObjectId)) throw new ArgumentNullException("Id cannot be empty.");
-            if (string.IsNullOrWhiteSpace(input.SObjectType)) throw new ArgumentNullException("Type cannot be empty.");
-
+            ValidationHandler.Run(input, connection);
             var client =
                 new RestClient(
-                    $"{input.Domain}/services/data/{input.ApiVersion}/sobjects/{input.SObjectType}/{input.SObjectId}");
+                    $"{connection.InstanceUrl}/services/data/{connection.ApiVersion}/sobjects/{input.SObjectType}/{input.SObjectId}");
             var request = new RestRequest("/", Method.Delete);
+            var accessToken = await ConnectionHandler.GetAccessToken(connection, cancellationToken);
+            request.AddHeader("Authorization", "Bearer " + accessToken);
 
-            switch (options.AuthenticationMethod)
-            {
-                case AuthenticationMethod.AccessToken:
-                    if (string.IsNullOrWhiteSpace(options.AccessToken))
-                        throw new ArgumentException(
-                            "Access token cannot be null when using Access Token authentication method");
-                    request.AddHeader("Authorization", "Bearer " + options.AccessToken);
-
-                    break;
-                case AuthenticationMethod.OAuth2WithPassword:
-                    accessToken = await GetAccessToken(options.AuthUrl, options.ClientID, options.ClientSecret,
-                        options.Username, options.Password + options.SecurityToken, cancellationToken);
-                    request.AddHeader("Authorization", "Bearer " + accessToken);
-
-                    break;
-            }
-
-            if (!(options.AuthenticationMethod is AuthenticationMethod.OAuth2WithPassword && options.ReturnAccessToken))
+            if (!ConnectionHandler.ShouldReturnToken(connection.AuthenticationMethod, connection.ReturnAccessToken))
                 accessToken = string.Empty;
+
             var response = await client.ExecuteAsync(request, cancellationToken);
             var content = JsonConvert.DeserializeObject<dynamic>(response.Content);
 
-            if (options.ThrowAnErrorIfNotFound && response.ErrorException.ToString()
+            if (connection.ThrowAnErrorIfNotFound && response.ErrorException.ToString()
                     .Equals(new HttpRequestException("Request failed with status code NotFound").ToString()))
                 throw new HttpRequestException("Target couldn't be found with given id.");
-
 
             return new Result(content, response.IsSuccessful, response.ErrorException, response.ErrorMessage,
                 accessToken);
@@ -79,30 +61,4 @@ public class Salesforce
             return Helpers.ErrorHandler.Handle(e);
         }
     }
-
-
-    #region HelperMethods
-
-    /// <summary>
-    /// Get OAuth2 access token.
-    /// </summary>
-    internal static async Task<string> GetAccessToken(string url, string clientId, string clientSecret, string username,
-        string passwordWithSecurityToken, CancellationToken cancellationToken)
-    {
-        var authClient = new RestClient(url);
-        var authRequest = new RestRequest("", Method.Post);
-        authRequest.AddHeader("Content-Type", "application/x-www-form-urlencoded");
-        authRequest.AddParameter("grant_type", "password");
-        authRequest.AddParameter("client_id", clientId);
-        authRequest.AddParameter("client_secret", clientSecret);
-        authRequest.AddParameter("username", username);
-        authRequest.AddParameter("password", passwordWithSecurityToken);
-        var authResponse = await authClient.ExecuteAsync(authRequest, cancellationToken);
-        string accessToken = JsonConvert.DeserializeObject<dynamic>(authResponse.Content).access_token;
-
-        return accessToken;
-    }
-
-    #endregion
-
 }